Include not the main lib into the Controller
